refactor: merge metadata and utils files
