+++ crates/ark-metadata/src/lib.rs
@@ -1,6 +1,5 @@
 mod cairo_string_parser;
 pub mod file_manager;
-mod metadata;
 pub mod metadata_manager;
 pub mod storage;
 pub mod types;
